Typesetter changes to ensure ruby 1.9 compatibility
This updates the typesetter class to use the ord method on each character in
the string. In 1.8.7 the conversion was automatic, but ord works in both 1.8.7
and 1.9.2. It also updates the spec_helper to the Rspec class instead of Spec
to update for later versions of Rspec. All tests still pass, except the 2
pending tests.